// app/react/edge/edge-jobs/ItemView/UpdateEdgeJobForm/types.ts
import { EdgeGroup } from '@/react/edge/edge-groups/types';
import { EnvironmentId } from '@/react/portainer/environments/types';

import { timeOptions } from '../../components/EdgeJobForm/RecurringFieldset';

export interface FormValues {
  name: string;
  recurring: boolean;
  edgeGroupIds: Array<EdgeGroup['Id']>;
  environmentIds: Array<EnvironmentId>;

  fileContent: string;

  cronMethod: 'basic' | 'advanced';
  dateTime: Date; // basic !recurring
  recurringOption: (typeof timeOptions)[number]['value']; // basic recurring
  cronExpression: string; // advanced
}